+++ cmd/goatcounter/reindex.go
@@ -98,7 +98,7 @@ func cmdReindex(f zli.Flags, ready chan<- struct{}, stop chan struct{}) error {
 
 		zlog.Config.SetDebug(debug)
 
-		db, ctx, err := connectDB(dbConnect, nil, false, false)
+		db, ctx, err := connectDB(dbConnect, []string{"pending"}, false, false)
 		if err != nil {
 			return err
 		}